+++ apps/example-app/app/devtools/ReactotronConfig.ts
@@ -3,7 +3,7 @@
  * free desktop app for inspecting and debugging your React Native app.
  * @see https://github.com/infinitered/reactotron
  */
-import { Platform, NativeModules } from "react-native"
+import { Platform, TurboModuleRegistry } from "react-native"
 
 import AsyncStorage from "@react-native-async-storage/async-storage"
 import { ArgType } from "reactotron-core-client"
@@ -29,7 +29,7 @@ const reactotron = Reactotron.configure({
     mst({
       /** ignore some chatty `mobx-state-tree` actions  */
       filter: (event) => /postProcessSnapshot|@APPLY_SNAPSHOT/.test(event.name) === false,
-    }),
+    })
   )
 
 if (Platform.OS !== "web") {
@@ -53,15 +53,23 @@ if (Platform.OS !== "web") {
  * NOTE: If you edit this file while running the app, you will need to do a full refresh
  * or else your custom commands won't be registered correctly.
  */
-reactotron.onCustomCommand({
-  title: "Show Dev Menu",
-  description: "Opens the React Native dev menu",
-  command: "showDevMenu",
-  handler: () => {
-    Reactotron.log("Showing React Native dev menu")
-    NativeModules.DevMenu.show()
-  },
-})
+
+/**
+ * This Platform.OS iOS restriction can be lifted in React Native 0.77
+ * The `DevMenu` module was missing in Android for the New Architecture
+ * See this PR for more details: https://github.com/facebook/react-native/pull/46723
+ */
+if (Platform.OS === "ios") {
+  reactotron.onCustomCommand({
+    title: "Show Dev Menu",
+    description: "Opens the React Native dev menu",
+    command: "showDevMenu",
+    handler: () => {
+      Reactotron.log("Showing React Native dev menu")
+      TurboModuleRegistry.get<{ show: () => void; getConstants: () => {} }>("DevMenu")?.show()
+    },
+  })
+}
 
 reactotron.onCustomCommand({
   title: "Reset Root Store",