+++ lib/reactotron-apisauce/CHANGELOG.md
@@ -2,6 +2,15 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+### [3.0.6](https://github.com/infinitered/reactotron/compare/reactotron-apisauce@3.0.5...reactotron-apisauce@3.0.6) (2025-01-07)
+
+
+### Bug Fixes
+
+* **package.json:** remove "react-native" export condition ([#1530](https://github.com/infinitered/reactotron/issues/1530)) ([47be69f](https://github.com/infinitered/reactotron/commit/47be69fae1e410a22907893a69474857a3e5a854))
+
+### [3.0.5](https://github.com/infinitered/reactotron/compare/reactotron-apisauce@3.0.4...reactotron-apisauce@3.0.5) (2024-08-26)
+
 ### [3.0.4](https://github.com/infinitered/reactotron/compare/reactotron-apisauce@3.0.3...reactotron-apisauce@3.0.4) (2024-02-14)
 
 ## [3.0.3](https://github.com/infinitered/reactotron/compare/reactotron-apisauce@3.0.2...reactotron-apisauce@3.0.3) (2024-02-14)