+++ lib/reactotron-react-js/CHANGELOG.md
@@ -2,6 +2,15 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+### [3.3.16](https://github.com/infinitered/reactotron/compare/reactotron-react-js@3.3.15...reactotron-react-js@3.3.16) (2025-01-07)
+
+
+### Bug Fixes
+
+* **package.json:** remove "react-native" export condition ([#1530](https://github.com/infinitered/reactotron/issues/1530)) ([47be69f](https://github.com/infinitered/reactotron/commit/47be69fae1e410a22907893a69474857a3e5a854))
+
+### [3.3.15](https://github.com/infinitered/reactotron/compare/reactotron-react-js@3.3.14...reactotron-react-js@3.3.15) (2024-08-26)
+
 ### [3.3.14](https://github.com/infinitered/reactotron/compare/reactotron-react-js@3.3.13...reactotron-react-js@3.3.14) (2024-04-04)
 
 ### [3.3.13](https://github.com/infinitered/reactotron/compare/reactotron-react-js@3.3.12...reactotron-react-js@3.3.13) (2024-02-14)