+++ lib/reactotron-react-native-mmkv/CHANGELOG.md
@@ -2,6 +2,15 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+### [0.2.8](https://github.com/infinitered/reactotron/compare/reactotron-react-native-mmkv@0.2.7...reactotron-react-native-mmkv@0.2.8) (2025-01-07)
+
+
+### Bug Fixes
+
+* **package.json:** remove "react-native" export condition ([#1530](https://github.com/infinitered/reactotron/issues/1530)) ([47be69f](https://github.com/infinitered/reactotron/commit/47be69fae1e410a22907893a69474857a3e5a854))
+
+### [0.2.7](https://github.com/infinitered/reactotron/compare/reactotron-react-native-mmkv@0.2.6...reactotron-react-native-mmkv@0.2.7) (2024-08-26)
+
 ### [0.2.6](https://github.com/infinitered/reactotron/compare/reactotron-react-native-mmkv@0.2.5...reactotron-react-native-mmkv@0.2.6) (2024-02-28)
 
 ### [0.2.5](https://github.com/infinitered/reactotron/compare/reactotron-react-native-mmkv@0.2.4...reactotron-react-native-mmkv@0.2.5) (2024-02-14)